rename `openInlangProject` to `loadProject`
simplifies communication and understanding. a user has the intent to _open_ a project. the sdk _loads_ a project.
