+++ inlang/source-code/ide-extension/src/main.ts
@@ -15,7 +15,7 @@ import { recommendation, isInWorkspaceRecommendation } from "./utilities/recomme
 import { linterDiagnostics } from "./diagnostics/linterDiagnostics.js"
 import { openInEditorCommand } from "./commands/openInEditor.js"
 import { editMessageCommand } from "./commands/editMessage.js"
-import { openInlangProject } from "@inlang/sdk"
+import { loadProject } from "@inlang/sdk"
 import { createFileSystemMapper } from "./utilities/createFileSystemMapper.js"
 import { _import } from "./utilities/import/_import.js"
 import { tryCatch } from "@inlang/result"
@@ -111,7 +111,7 @@ async function main(args: {
 	)
 
 	const { data: inlang, error } = await tryCatch(() =>
-		openInlangProject({
+		loadProject({
 			settingsFilePath: closestProjectFilePathUri.fsPath,
 			nodeishFs: createFileSystemMapper(workspaceFolder.uri.fsPath),
 			_import: _import(workspaceFolder.uri.fsPath),
